backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-generativeai
pinecone
pydantic
python-multipart
pypdf
python-docx
